# docs/api/runtime-v3/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ADL Runtime v3 Core API",
    "version": "1.0.0",
    "summary": "Control, status, telemetry, and agent communication",
    "description": "Use this API to inspect runtime health, read telemetry, send authenticated control commands, and exchange ACIP messages."
  },
  "servers": [
    {
      "url": "{publicBaseUrl}",
      "description": "Runtime HTTPS base URL.",
      "variables": {
        "publicBaseUrl": {
          "default": "https://localhost:20997",
          "description": "Base URL for the runtime instance."
        }
      }
    }
  ],
...
  "paths": {
...
    "/v1/acip/ws": {
      "get": {
        "operationId": "openAcipWebSocket",
        "tags": [
          "acip"
        ],
        "summary": "Open the authenticated bidirectional ACIP WebSocket carrier",
        "description": "Upgrades to WebSocket after bearer authentication. The server sends an authenticated session frame, accepts bounded ACIP Protobuf binary envelopes, dispatches each envelope through canonical ingress to the runtime service named by its route, and returns a structured completion or rejection frame.",
        "security": [
          {
            "runtimeBearer": []
          }
        ],
        "responses": {
          "101": {
            "description": "WebSocket upgrade accepted."
          },
          "401": {
            "$ref": "#/components/responses/RuntimeAuthenticationFailed"
          }
        },
        "x-adl-websocket": {
          "scheme": "wss",
          "maxFrameBytes": 65536,
          "authentication": "http_bearer_upgrade_header",
          "serverFirstFrame": {
            "$ref": "#/components/schemas/AcipAuthenticatedFrame"
          },
          "clientFrames": [
            {
              "binarySchema": "adl.csm.acip_carrier.protobuf_envelope.v1",
              "maxBytes": 65536
            }
          ],
          "serverFrames": [
            {
              "$ref": "#/components/schemas/AcipDispatchResult"
            }
          ],
          "policyCloseReasons": [
            "credential_revoked",
            "binary_acip_frame_required"
          ],
          "bidirectional": true
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "runtimeBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token configured outside code through Runtime v3 init/environment."
      },
      "runtimeSignedCommand": {
        "type": "mutualTLS",
        "description": "Runtime transport is TLS/Rustls. Command authorization is additionally enforced by the Ed25519 signature fields in SignedControlCommand."
      }
    },
    "responses": {
      "RuntimeAuthenticationFailed": {
        "description": "Missing, malformed, incorrect, or unavailable Runtime API bearer credential.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ControlError"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "AcipAuthenticatedFrame": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "event",
          "path",
          "bidirectional"
        ],
        "properties": {
          "schema": {
            "const": "adl.csm.runtime_api.wss_session.v1"
          },
          "event": {
            "const": "authenticated"
          },
          "path": {
            "const": "/v1/acip/ws"
          },
          "bidirectional": {
            "const": true
          }
        }
      },
      "AcipDispatchResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schema", "status", "sequence_reserved"],
        "properties": {
          "schema": {
            "const": "adl.csm.acip_carrier.websocket_frame.v1"
          },
          "status": {
            "enum": ["completed", "rejected"]
          },
          "message_id": {
            "type": "string"
          },
          "accepted_sequence": {
            "type": "integer",
            "minimum": 0
          },
          "result_hash": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "sequence_reserved": {
            "type": "boolean"
          }
        }
      },
...
      "ControlError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "code"
        ],
        "properties": {
          "schema": {
            "const": "adl.runtime.control_error.v1"
          },
          "code": {
            "enum": [
              "authentication_failed",
              "unauthorized",
              "idempotency_conflict",
              "temporarily_unavailable",
              "stale_runtime_instance",
              "invalid_request"
            ]
          }
        }
      }
    }
  },
...
}